[chore][pkg/stanza] Cleanup debug input operator files
