+++ pkg/stanza/operator/input/generate/input_test.go
