+++ pkg/stanza/operator/input/stdin/input_test.go
